backend/requirements.txt
flask>=2.3.0
python-dotenv>=1.0.0
pdfminer.six>=20221105
python-docx>=0.8.11
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
google-generativeai>=0.7.0
flask-login>=0.6.3
werkzeug>=2.3.0
